cmake: tweak static/import lib name collision avoidance logic
The collision issue affects (typically) MSVC, when building both shared
and static libssh2 in one go.

Ref: https://stackoverflow.com/questions/2140129/what-is-proper-naming-convention-for-msvc-dlls-static-libraries-and-import-libr

Initially we handled this by appending the `_imp` suffix to the import
library filename. This is how curl tackles this, but on a second look,
this solution seem to be accidental and has no widespread use.

It seems more widely accepted to use the '_static' suffix for the static
library. This patch implements this.

(MinGW, Cygwin and unixy platforms are not affected by this issue.)

Follow-up to 4e2580628dd1f8dc51ac65ac747ebcf0e93fa3d1

Cherry-picked from #1036