use built-in `_WIN32` macro to detect Windows
`_WIN32` is defined by the compiler, while `WIN32` is defined by
the Windows headers / build env, or we had to take care of it. The
agreement seems to be that `_WIN32` is the preferred practice here.

Minor downside is that CMake uses `WIN32` and it was also adopted
in `Makefile.mk`.

In public libssh2 headers we stick with accepting either `_WIN32` or
`WIN32` and define our own namespaced `LIBSSH2_WIN32` based on them.

grepping for `WIN32` remains useful to detect Windows-specific code.

Closes #xxxxx