mbedtls: improve disabling `-Wredundant-decls`
Disable these warnings specifically for the mbedTLS public headers
and leave it on for the the rest of the code. This also fixes this
issue for autotools. Previous solution was globally disabling this
warning for the whole code when using mbedTLS and only with CMake.

Follow-up to 7ecc309cd10454c54814b478c4f85d0041da6721 #1224

Closes #1226