Implement posix-rename@openssh.com
This commit adds a new function, libssh2_sftp_posix_rename_ex,
which implements the posix-rename@openssh.com extension.

If the server does not support this extension, the function will
return LIBSSH2_FX_OP_UNSUPPORTED and it will be up to the user to
recover, possibly by calling libssh2_sftp_rename.