mbedtls: convert to native pkg-config detection
Hm, when using `mbedtls`, pkg-config picks up the mbedtls as the only
lib, and it fails while linking. Seems like the absolute path is meant
to point to the shared lib, while the bare lib to the static? Though
there is a separate STATIC version returned of each result variable.