cmake: always run the tests even when static libs are disabled
The BUILD_STATIC_LIBS option is intended for the use case where users do
not wish to build and install a static lib because it is not one of the
artifacts they are interested in. The option should not be used to
disable running the testsuite.

Instead, when tests are enabled, define the target. Just don't export or
install it. Build systems build lots of things that aren't deliverable
artifacts, and this becomes just one more of them.