ci/appveyor: stop running tests
Jobs consistently fail to connect to the test server (run in GHA) since
2024-Aug-29:
https://ci.appveyor.com/project/libssh2org/libssh2/builds/50498393

There was an earlier phase of failures one month before that, that got
fixed by increasing the wait for the server in
bf3af90b3f1bb14cf452df7a8eb55cc9088f3e7f.

Let's make that wait 10 minues (was 5), in case it helps, though I tried
it earlier to no avail. Such long wait doesn't seem acceptable if it
works, and a sign of issues lurking somewhere.

Thus, skip running tests in AppVeyor CI. This leaves the Windows
platform with no tests at all.

Except in the last job to serve as a canary in case the tests would
"fix themselves", and to leave a signal that we have problem.