cmake: only run static tests when BUILD_TESTING is passed on the command line
The default value of BUILD_TESTING is "on", which means that all users
generate a test plan, even if they don't run tests. It is my belief that
this is not a problematic occurrence, but due to concerns I have
acquiesced and implemented this modification.

It works for our use cases because we do, after all, pass both
-DBUILD_TESTING and -DBUILD_STATIC_LIBS=OFF. :) As such, we still get
all tests that we would get if BUILD_STATIC_LIBS were enabled. People
relying on the default option values will either run a smaller test plan
or have to adapt their build commands.

This therefore means that users who only use -DBUILD_STATIC_LIBS=OFF do
not generate any build rules to create an internal convenience library,
since no test programs depend on one. Exported targets do not change at
all, since no exported target has ever been created anyway (that is
controlled by -DBUILD_STATIC_LIBS=ON, which is unrelated to the test
target rework), but that's not the point.