cmake: always define the STATIC_LIB_SUFFIX when clashes could happen
When using the relevant compiler toolchain and building shared
libraries, it is not possible to name the static library the same as the
shared library. This rule applies regardless of whether static libraries
are built, since it is not possible to name a static library the same as
a shared library even if you solve this by not, in fact, building a
static library at all.

So, it is not necessary to check whether static libraries are requested
before setting this variable. If static libraries aren't requested, the
variable will simply do nothing as it is never consulted.

This makes the logic shorter and simpler, and also makes it less brittle
when a later patch modifies the condition for invoking
`add_library(... STATIC)`.