libssh2.h: Add note about release versions to LIBSSH2_VERSION
As noted in #771 and #1473, the blessed and signed releases are on
libssh2.org and any other archives, like those generated by GitHub
are just a collection of files from the tag and are not the official
releases.

This adds a note, which already exists for LIBSSH2_TIMESTAMP, to
LIBSSH2_VERSION, to clarify this fact.