Find: set case-sensitive `<find-modulename>_FOUND` variables
E.g. `MbedTLS_FOUND`.

It's necessary because `find_dependency()` (as of CMake 3.31.5)
does check for this variable when deciding if a Find module
had found the dependency successfully. It does not check for
the `<uppercase-module-name>_FOUND` name.

https://github.com/Kitware/CMake/blob/eacc7287297267203fd65be1ca1daaaf62ebca37/Modules/CMakeFindDependencyMacro.cmake#L93-L97

Since find_package_handle_standard_args() is setting both of
these `_FOUND` variables, make sure we set it on the pkg-config
codepath which doesn't call find_package_handle_standard_args().