Buffer overread vulnerability fix
If Address Sanitization (asan) is enabled when building libssh2 and any encapsulating code, attempts to call the vulnerable functions using mismatched lengths will cause the application to panic and crash. If asan is not enabled during the build process, the application will continue running with the malicious buffer.
The affected functions send the data to the connected SSH server, leaking memory from the server running libssh2. Due to the nature of libssh2, and code using the library, there is a high probability of sensitive information such as authentication material being leaked to the remote server.