Update FindMbedTLS.cmake
This module is a compatibility layer for the library version below 3.6.0,
where full support of cmake configuration modules appeared. It is recommended
to use the library 3.6 and higher, for which this module is redundant.

The module provides the same targets as the MbedTLS 3.6 modules, due to which
the migration process to 3.6 will be very simple - just delete the local copy
of this file and reinitialize the project.

Signed-off-by: Azamat H. Hackimov <azamat.hackimov@gmail.com>