cmake: do not de-duplicate `LIBSSH2_PC_LIBS_PRIVATE`
And the libs listed in `Libs.private` in `libssh2.pc` with it.

The order and repetition might be significant for linkers that rely on
strict lib order and unable to resolve symbols without it. Such linker
is binutils `ld`.

This also syncs this snippet with curl.

De-duplication was originally introduced to avoid a repeat `-lz`. It
means after this patch a repeat `-lz` is back again. (with
a zlib-enabled OpenSSL or wolfSSL build.)

One safe way to de-duplicate would be to remove dupes when they appear
right next to each other. This is an excercise for another PR, because
cmake doesn't have such built-in function. Meanwhile, such duplicates
don't cause harm, and only make the output less clean-looking and
introduce potential differences between `libssh2.pc` generated by cmake
and autotools. (those aren't necessary identical anyway, yet.)

Follow-up to 6464301820a9ca4a56c5f02717430bbd4150c7b2 #1131